+++ .gitignore
@@ -1,2 +1,39 @@
 /U4/TCP-Chat_s4jeheid/.idea/
 /U4/TCP-Chat_s4jeheid/Server.jar
+/out/production/UDP-TCP-Chat/.idea/.gitignore
+/out/production/UDP-TCP-Chat/Java - netcat TCP/.idea/.gitignore
+/out/production/UDP-TCP-Chat/.gitignore
+/U4/TCP-Chat-standalone-04/.idea/.gitignore
+/out/production/UDP-TCP-Chat/.idea/libraries/artifacts.xml
+/out/production/UDP-TCP-Chat/U4/Aufgabe_2_CIDR.txt
+/U4/TCP-Chat-standalone-04/.idea/encodings.xml
+/out/production/UDP-TCP-Chat/U4/TCP-Chat_s4jeheid/FP_S24_ueb09.iml
+/out/production/UDP-TCP-Chat/U4/FP_S24_ueb09_s4jeheid.zip
+/out/production/UDP-TCP-Chat/U4/IPv4-Header.png
+/out/production/UDP-TCP-Chat/Java - netcat TCP/Java - netcat TCP.iml
+/out/production/UDP-TCP-Chat/Java - netcat UDP/Java - netcat UDP.iml
+/out/production/UDP-TCP-Chat/Java - netcat TCP/out/artifacts/Java-netcat_TCP.jar
+/out/production/UDP-TCP-Chat/Java - netcat UDP/out/artifacts/Java-netcat_UDP.jar
+/out/production/UDP-TCP-Chat/.idea/libraries/Java_netcat_TCP.xml
+/out/production/UDP-TCP-Chat/Java - netcat TCP/.idea/libraries/Java_netcat_TCP.xml
+/out/production/UDP-TCP-Chat/Manchester-Code_A4.png
+/out/production/UDP-TCP-Chat/Manchester-Code_wavedrom.json
+/out/production/UDP-TCP-Chat/Java - netcat TCP/src/META-INF/MANIFEST.MF
+/out/production/UDP-TCP-Chat/Java - netcat UDP/META-INF/MANIFEST.MF
+/out/production/UDP-TCP-Chat/Java - netcat UDP/src/META-INF/MANIFEST.MF
+/out/production/UDP-TCP-Chat/.idea/misc.xml
+/out/production/UDP-TCP-Chat/Java - netcat TCP/.idea/misc.xml
+/U4/TCP-Chat-standalone-04/.idea/misc.xml
+/out/production/UDP-TCP-Chat/.idea/modules.xml
+/out/production/UDP-TCP-Chat/Java - netcat TCP/.idea/modules.xml
+/out/production/UDP-TCP-Chat/Java - netcat UDP/out/artifacts/netcatUDP.jar
+/out/production/UDP-TCP-Chat/U4/TCP-Chat_s4jeheid/pom.xml
+/out/production/UDP-TCP-Chat/README.md
+/out/production/UDP-TCP-Chat/U4/TCP-Header.png
+/out/production/UDP-TCP-Chat/U4/TCP-Header-colorised.png
+/out/production/UDP-TCP-Chat/U4/UDP-Header.png
+/out/production/UDP-TCP-Chat/U4/UDP-Header-colorized.png
+/out/production/UDP-TCP-Chat/UDP-TCP-Chat.iml
+/out/production/UDP-TCP-Chat/.idea/vcs.xml
+/out/production/UDP-TCP-Chat/Java - netcat TCP/.idea/vcs.xml
+/U4/TCP-Chat-standalone-04/.idea/vcs.xml
