--- U4/TCP-Chat_s4jeheid/src/main/java/de/unitrier/st/fp/s24/ueb09/Main.java
@@ -1,16 +1,0 @@
-package de.unitrier.st.fp.s24.ueb09;
-
-import javafx.application.Application;
-import javafx.stage.Stage;
-
-public class Main extends Application{
-
-    private final int NUMBER_OF_WINDOWS = 3;
-
-    @Override
-    public void start(Stage stage) throws Exception {
-        for (int i = 0; i < NUMBER_OF_WINDOWS; i++) {
-            new ChatWindowClient().createChatWindow();
-        }
-    }
-}
